+++ src/components/todo-list.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+import TodoListItem from './todo-list-item';
+
+const TodoList = () => {
+  return (
+    <ul>
+      <li><TodoListItem /></li>
+      <li><TodoListItem /></li>    
+    </ul>
+  );
+};
+
+export default TodoList;
