+++ src/index.js
@@ -1,30 +1,9 @@
 import React from 'react';
 import ReactDom from 'react-dom';
 
-const TodoList = () => {
-  const items = ['Learn React', 'Build Awesome App'];
-  return (
-    <ul>
-      <li>{ items[0] }</li>
-      <li>{ items[1] }</li>    
-    </ul>
-  );
-};
-
-const AppHeader = () => {
-  return <h1>My Todo List</h1>;
-};
-
-const SearchPanel = () => {
-  const searchText = 'Type here to search';
-  const searchStyle = {
-    fontSize: '20px'
-  };
-
-  return <input
-    style={searchStyle} 
-    placeholder={searchText}/>;
-};
+import AppHeader from './components/app-header';
+import SearchPanel from './components/search-panel';
+import TodoList from './components/todo-list'
 
 const App = () => {
   return (
